[PCX-1472] Extract view management logic
